requirements.txt
numpy==1.26.4
pandas==2.2.3
tqdm==4.66.6
konlpy==0.6.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
faiss-gpu==1.7.2
langchain==0.3.4
langchain_community==0.3.3
sentence-transformers==3.2.1